eval broker: shed blocked evals older than acknowledged eval
When an evaluation is acknowledged by a scheduler, the resulting plan is
guaranteed to cover up to the `ModifyIndex` set by the worker based on the
most recent evaluation for that job in the state store. At that point, we no
longer need to retain blocked evaluations in the broker that are older than that
index.

Move these stale evals into a canceled set. When the `Eval.Ack` RPC returns from
the eval broker it will retrieve a batch of canceable evals to write to
raft. This paces the cancelations limited by how frequently the schedulers are
acknowledging evals; this should reduce the risk of cancelations from
overwhelming raft relative to scheduler progress.